pyproject.toml
[project]
name = "polyarb"
version = "0.1.0"
description = "Prediction market aggregator & arbitrage detector"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "websockets>=14.0",
    "apscheduler>=3.10.0",
    "redis[hiredis]>=5.2.0",
    "scikit-learn>=1.6.0",
    "rapidfuzz>=3.11.0",
    "kalshi-python-sync>=2.1.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
